Refactor integration module into modular, composable components
